Fix code

+++ src/main/java/com/github/gilbertotcc/himitsubako/Factorial.java
@@ -1,5 +1,6 @@
 package com.github.gilbertotcc.himitsubako;
 
+import java.math.BigInteger;
 import java.util.Collections;
 import java.util.List;
 
@@ -7,12 +8,12 @@ public class Factorial {
 
     private Factorial() {}
 
-    public static long factorialOf(final long number) {
+    public static BigInteger factorialOf(final int number) {
         // TODO
-        return number;
+        return BigInteger.valueOf(number);
     }
 
-    public static List<Long> factorialOf(final List<Long> numbers) {
+    public static List<BigInteger> factorialOf(final List<Integer> numbers) {
         // TODO
         return Collections.emptyList();
     }
